Reorganise file locations and add file to demo pipeline
