+++ common_components/data_validator/general_validators/format_validator.py
@@ -3,8 +3,8 @@ import numpy as np
 import json
 from typing import Dict, Any
 
-from common_components.validation_handler import ValidationHandler
-from validator_logger import ValidatorLogger
+from common_components.data_validator.validation_handler import ValidationHandler
+from common_components.data_validator.validator_logger import ValidatorLogger
 
 
 class FormatValidator(ValidationHandler):
