+++ common_components/data_validator/general_validators/not_empty_validator.py
@@ -1,8 +1,8 @@
 import pandas as pd
 import numpy as np
 from typing import Dict, Any
-from common_components.validation_handler import ValidationHandler
-from validator_logger import ValidatorLogger
+from common_components.data_validator.validation_handler import ValidationHandler
+from common_components.data_validator.validator_logger import ValidatorLogger
 
 class NotEmptyValidator(ValidationHandler):
     """
