+++ common_components/data_validator/text_validator/only_string_validator.py
@@ -1,6 +1,6 @@
 from typing import Dict, Any
-from common_components.validation_handler import ValidationHandler
-from validator_logger import ValidatorLogger
+from common_components.data_validator.validation_handler import ValidationHandler
+from common_components.data_validator.validator_logger import ValidatorLogger
 
 class OnlyStringValidator(ValidationHandler):
     """
